requirements.txt
abstract_singleton==1.0.1
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
auth0-python==4.4.2
auto_gpt_plugin_template==0.0.3
autoflake==2.2.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.2
bidict==0.22.1
black==23.3.0
bleach==6.0.0
blinker==1.6.1
blis==0.7.10
Brotli==1.1.0
cachetools==5.3.0
catalogue==2.0.9
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chainlit==0.6.402
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
confection==0.1.3
contourpy==1.0.7
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.8
Cython==3.0.2
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.0.7
distlib==0.3.6
distro==1.8.0
docker==6.1.3
docopt==0.6.2
docx==0.2.4
docx2txt==0.8
duckduckgo-search==3.8.5
EbookLib==0.18
entrypoints==0.4
executing==1.2.0
fastapi==0.99.1
fastapi-socketio==0.0.10
fastjsonschema==2.16.2
filelock==3.12.2
filetype==1.2.0
Flask==2.2.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.0
ftfy==6.1.1
futu-api==7.3.3508
gitdb==4.0.10
GitPython==3.1.31
googleapis-common-protos==1.60.0
-e git+https://github.com/AntonOsika/gpt-engineer@b2186ad7ebc4fd2b78f4aafe4b3a5bfc72facbd7#egg=gpt_engineer
grpcio==1.58.0
gTTS==2.3.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.17.0rc0
hyperframe==6.0.1
identify==2.5.24
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langchain==0.0.285
langcodes==3.3.0
langsmith==0.0.35
Lazify==0.4.0
llama-index==0.5.9
lxml==4.9.2
macholib==1.16.2
mariadb==1.1.6
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
multidict==6.0.4
murmurhash==1.0.10
mypy==1.3.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
mysqlclient==2.2.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nodeenv==1.8.0
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.5
numpy==1.24.2
openai==1.3.6
openapi-python-client==0.15.2
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
orjson==3.9.7
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.2
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
platformdirs==3.8.1
playsound==1.3.0
pluggy==1.2.0
pre-commit==3.3.3
preshed==3.0.9
prisma==0.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
pyflakes==3.1.0
Pygments==2.14.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyJWT==2.8.0
pylatexenc==2.10
Pympler==1.0.1
PyMySQL==1.0.2
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
python-engineio==4.7.0
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-socketio==5.9.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.3.23
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.3
rudder-sdk-python==2.0.2
ruff==0.0.272
scikit-learn==1.2.1
scipy==1.10.1
semver==3.0.0
Send2Trash==1.8.0
shellingham==1.5.3
simplejson==3.19.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.47
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
streamlit==1.20.0
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.1.0
tiktoken==0.3.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions @ file:///private/tmp/python-typing-extensions-20230702-5533-jv3t41/typing_extensions-4.7.1/dist/typing_extensions-4.7.1-py3-none-any.whl
tzdata==2023.3
tzlocal==4.3
uptrace==1.19.0
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.23.2
validators==0.20.0
virtualenv==20.23.1
wasabi==1.1.2
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
yarg==0.1.9
yarl==1.8.2
youtube-transcript-api==0.6.1
zipp==3.15.0
